// rl3x1/src/app/back-office/Orders/Models/Orders.model.ts
export class OrdersModel {
    order_Id: number ;
    order_number: string ;
    vendor_number: string ;
    req_by_date: Date ;
    price_year: string ;
    season_code_id: string ;
    product_number: string ;
    category_code_id: string ;
    grade_code_id: string ;
    subcategory_code_id: string ;
    sex_code: string ;
    hstariff_id: string ;
    country_of_origin_id: string ;
    prim_season_code: string ;
    product_sku: string ;
    product_qty: string ;
    product_weight: string ;
    size_code_id: string ;
    sort_order: string ;
    color_description: string ;
    base_color: string ;
    hex_value: string ;
    createdby: number ;
    created_date: Date ;
    updatedby: number ;
    updated_date: Date ;
    IsActive: boolean;
    UserID: number;
    CaseNo: string;
    isActive: string;
}